backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
openai==1.3.7
pydantic==2.5.0
sqlalchemy==2.0.23
google-generativeai>=0.3.2

# Machine Learning
xgboost==2.0.3
torch --index-url https://download.pytorch.org/whl/cpu
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
yfinance>=0.2.50

# Technical Analysis
# pandas-ta  # Installed separately in Dockerfile to avoid conflicts
joblib
pytz
beautifulsoup4
lxml
